Require phpspec/prophecy as dev dependency
https://github.com/sebastianbergmann/phpunit/issues/5033 changed phpunit
to no longer depend on phpspec/prophecy. This results in the need to add
a development dependency for phpspec/prophecy. Doing so results in a
warning, since PHPUnit\Framework\TestCase::prophesize() is deprecated
and will be removed in PHPUnit 10. While we could use the trait provided
by phpspec/prophecy-phpunit instead, it would break PHP < 7.3 support.

* Closes #50